pyproject.toml
[project]
name = "django-sample-components"
version = "0.1.7"
description = "A handy collection of reusable Django UI components, allowing me to share and integrate elements efficiently across various projects."
authors = [
    {name = "Gustavo Rizzo S M de Albuquerque",email = "grizzo.albu@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python =  ">=3.14,<4.0"
dependencies = [
    "django (>=5.2.0,<7.0.0)",
    "django-htmx (>=1.27.0,<2.0.0)",
    "django-crispy-forms (>=2.3,<3.0.0)",
    "crispy-bootstrap5 (>=2024.10,<2025.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
mypy = "^1.18.2"
taskipy = "^1.14.1"
ruff = "^0.14.3"
django-debug-toolbar = "^6.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
exclude = [
    "migrations",
    "__pycache__",
    "build",
    "dist"
]

[tool.ruff.lint]
select = ["F", "E", "W", "I", "N", "UP", "B"]

[tool.taskipy.tasks]
lint = { cmd = "ruff check .", help = "runs the linter" }
lint-fix = { cmd = "ruff check . --fix", help = "runs the linter and tries to fix issues" }
test = { cmd = "cd demo_project && python manage.py test", help = "runs the test suite" }
run-demo = { cmd = "cd demo_project && python manage.py runserver", help = "runs the demo project server" }
collectstatic = { cmd = "cd demo_project && python manage.py collectstatic --noinput", help = "collects static files" }
